+++ src/agent/modelHandlers/modelHandlerAnthropic.ts
@@ -40,7 +40,7 @@ import {
 import { ToolState } from '../ToolState';
 import { MediaEntry } from '../mediaTypes';
 import { AgentStateRound } from '../AgentState';
-import { K_SLICE } from '../../utils/config';
+import { K_SLICE } from '../../utils/config/constants';
 import { objectToLogString } from '../../utils/text/stringUtils';
 import { calculateTokenPrice } from '../../utils/priceUtils';
