+++ src/logger/AgentLogger.ts
@@ -4,7 +4,7 @@
 // Local imports - log
 import * as logger from './logUtils';
 import { sleep } from '../utils/helpers';
-import { SHORT_SLEEP_MS } from '../utils/config';
+import { SHORT_SLEEP_MS } from '../utils/config/constants';
 
 /**
  * Encapsulates logging functionality for agents with a dedicated channel.
