+++ src/utils/text/repetitionUtils.ts
@@ -10,7 +10,7 @@ import * as logger from '../../logger/logUtils';
 import {
   REPETITION_DETECTION_THRESHOLD,
   REPETITION_PREVIEW_LENGTH,
-} from '../config';
+} from '../config/constants';
 
 const CHANNEL = 'repetitionUtils';
 logger.initialize(CHANNEL);
